Landing Page (Prodigy Internship Task 1)
The iBook Project 01 is the first task in the Prodigy Web Development Internship. The goal of this project was to design and implement a responsive and visually appealing landing page using HTML and CSS. This project serves as an introduction to front-end development and demonstrates the ability to structure and style a modern web interface.
 Technologies Used:
HTML5: For semantic page structure.
CSS3: For styling, layout, and responsiveness.
Media Queries: To ensure responsiveness across different screen sizes.
 JavaScript: For animations or interactive elements
Conclusion:
The iBook Landing Page project served as a foundational exercise in web development. It reinforced the use of semantic HTML, clean CSS, and responsive design techniques. This task demonstrates readiness for more advanced front-end tasks in the internship